requirements.txt
h5py==2.10.0
numpy==1.23.5
onnxruntime==1.14.1
parallel_wavegan==0.6.1
PyYAML==6.0
PyYAML==6.0.1
soundfile==0.12.1
torch==1.12.1
tqdm==4.64.1
